requirements.txt
pbkdf2
pycrypto